Improve float/double writing logic
The logic leverages some internal HHVM functions
to avoid calling the slower `pack` function.

This is a 3% perf improvement on the existing
benchmark, 7% improvement on a float heavy benchmark.

Unfortunately we have to settle for a slightly
less readable variant of the code due to a bug in
the 4.153.1 (LTS) version of HHVM.

Also we seem to be hitting some HHVM issue as we
need to disable 2 typechecker errors when the
construct correctly turned into working bytecode.