Rework the benchmarking logic (#16)
This PR forces testing in repo auth mode. This is because HHVM is
optimized to run in this mode.
So measuring performance against other mode would introduce distortion
in the measurements.

This PR also improves measurements by:
1. Warming up HHVM to ensure consistent measurements. The first run is
usually an outlier as
there are some one-off initializations in the VM that would be included
in it.
2. Compute the average and standard deviation after the runs so those
can be easily measured when optimizing.

To be able to do 2, the infinite mode was removed, replaced by a large
loop count instead.