Add GitHub Action to run tests against master/PRs
The GitHub Action will automatically run tests on PRs moving
forward. For simplicity, this uses Ubuntu 18.04 and HHVM 4.164.0
as those were the closest to what we normally use. We could
update values in a future PR once we've confirmed that the
CI works.

The CI catched some type checker errors I introduced in
https://github.com/slackhq/proto-hack/pull/16 that is fixed
as part of this PR :facepalm: