Fix the Ubuntu CI
The `ubuntu-18.04` image has been unsupported for more than a
year and there are no workers on GitHub that runs it. This means
that our CI just waits for one worker and times out.

The solution is to switch to a custom Docker image based on the
HHVM public images and move all the tests run into the new image.
To avoid spinning multiple containers (and recomputing some of the
cached bazel actions), all tests were moved into a new script
`ci/run_tests.sh`.