Implement support for proto3 optional
proto3 optional requires opting into the
experimental feature in our code generator.

proto3 optional is implemented as a one-off
synthetic field so we must treat those as normal
objects in our codegenerator. You can find the
instructions on how to handle those in the upstream
protobuf documentation:
https://github.com/protocolbuffers/protobuf/blob/main/docs/implementing_proto3_presence.md

The PR adds tests for the logic for both binary
and JSON decoding. We are missing reflection
support but that seemed reasonable for now.