Inline the binary protos (#25)
protoc-gen-hack/plugin.go was generating the gzipped binary
ProtoDescriptorProto and returning it to protoc so we could save it to a
file. protoc was warning us about having those file's content being
invalid UTF-8.

Unfortunately newer protoc binary reject files with such a ayload. We do
want to keep the binary FileDescriptorProto logic for servers that can
send the binary descriptor to their clients but we can inline it in each
generated file.

While we don't need to gzip the content anymore, I kept this logic as we
were generating large binary strings in Hacklang.