Implement hack_namespace option for hack codegen (#31)
Modify proto-hack codegen to support a custom namespace for generated
hack code.
A new option called `hack_namespace` can be added to .proto files as a
compiler hint. Eg. `option (hack_namespace) = "fizzy\\bazzy";. When the
hack_namespace option is set, the generated hack classes will be placed
in the specified namespace.

### Change details:
- The namespace at the top of generated class is changed to custom hack
namespace if available.
- If a file has `hack_namespace`, all references to its fields in other
proto classes will use custom namespace prefix.
When running the codegen, the first phase is to build a namespace tree
for protobuf file. During this phase, a new attribute `cFqn` will also
be populated. This attribute is the fully qualified name of the field
prefixed with custom namespace. When generating output classes, `cFqn`
will be used to refer to fields instead of `fqn`, if it is available.